+++ src/mimic_macros/src/imp/enum_value.rs
@@ -23,7 +23,7 @@ pub fn enum_value(node: &EnumValue, t: Trait) -> TokenStream {
 
     // quote
     let q = quote! {
-        fn value(&self) -> i64 {
+        fn value(&self) -> i32 {
             match self {
                 #inner
             }
